+++ CHANGES.md
@@ -18,6 +18,7 @@ Changes in 12.6
 - CSS themes reduce bottom buttons height in stylesheet.css & zresponsive.css
 - Display short month and year labels instead of short month only in CategoryBreakdownTime.php
 - SQL order Transcript Grades by Course Title in Transcripts.fnc.php
+- Add Administrator option to Visible To User Profiles in Resources.fnc.php, rosariosis.sql & rosariosis_mysql.sql
 
 Changes in 12.5
 ---------------